Split <ViewerNavigation/> into component and container.
